Refactor generated rbs spy testing
