+++ gems/smithy/spec/spec_helper.rb
@@ -8,6 +8,7 @@ require 'stringio'
 require 'smithy'
 
 require_relative 'support/be_in_documentation_matcher'
+require_relative 'support/rbs_spy_test'
 
 module SpecHelper
   class << self
@@ -18,6 +19,7 @@ module SpecHelper
     #  `:schema`, `:client`, or `:server`.
     # @param [Hash] options Additional options to pass to the generator.
     # @option options [String] :fixture The name of the fixture to load.
+    # @option options [Boolean] :rbs_test (true) Enable RBS spy testing when the SMITHY_RUBY_RBS_TEST is set.
     # @return [String] The path to the directory where the generated code was
     #  written to.
     def generate(modules, type, options = {})
@@ -29,7 +31,7 @@ module SpecHelper
       $LOAD_PATH << ("#{sdk_dir}/lib")
       require plan.gem_name
 
-      setup_rbs_spytest(modules, sdk_dir) if options.fetch(:rbs_test, ENV.fetch('SMITHY_RUBY_RBS_TEST', nil))
+      RbsSpyTest.setup(modules, sdk_dir) if options.fetch(:rbs_test, true) && ENV.fetch('SMITHY_RUBY_RBS_TEST', nil)
       sdk_dir
     rescue StandardError => e
       cleanup(modules, sdk_dir) if sdk_dir
@@ -55,18 +57,6 @@ module SpecHelper
       Object.send(:remove_const, module_names.first)
     end
 
-    def load_rbs_environment(sdk_dir, load_collection: true)
-      require 'rbs'
-
-      loader = RBS::EnvironmentLoader.new(core_root: RBS::EnvironmentLoader::DEFAULT_CORE_ROOT)
-      loader.add(path: Pathname(File.join(__dir__.to_s, '../../smithy-client/sig')))
-      loader.add(path: Pathname(File.join(sdk_dir, '/sig')))
-
-      load_collection(loader) if load_collection
-
-      RBS::Environment.from_loader(loader).resolve_type_names
-    end
-
     private
 
     def load_model(modules, options)
@@ -83,68 +73,6 @@ module SpecHelper
         quiet: ENV.fetch('SMITHY_RUBY_QUIET', 'true') == 'true'
       }
       Smithy::Plan.new(model, type, plan_options)
-    end
-
-    def load_collection(loader)
-      collection_config_path = RBS::Collection::Config.find_config_path
-      lock_path = RBS::Collection::Config.to_lockfile_path(collection_config_path)
-      if lock_path.file?
-        lock = RBS::Collection::Config::Lockfile.from_lockfile(
-          lockfile_path: lock_path,
-          data: YAML.load_file(lock_path.to_s)
-        )
-      end
-      raise 'Missing RBS collection, ensure you have `rbs collection install`' unless lock
-
-      loader.add_collection(lock)
-    end
-
-    def to_absolute_typename(type_name)
-      RBS::Factory.new.type_name(type_name).absolute!
-    end
-
-    def classes_to_spy(mod, env, visited = Set.new)
-      classes = []
-      visited << mod
-      mod.constants.each do |c|
-        sub_mod = mod.const_get(c)
-        # module includes classes
-        next if !sub_mod.is_a?(Module) || visited.include?(sub_mod) || sub_mod == ::BasicObject
-
-        rbs_name = to_absolute_typename(sub_mod.name)
-        if env.module_name?(rbs_name)
-          classes << sub_mod
-          classes += classes_to_spy(sub_mod, env, visited)
-        end
-      end
-      classes
-    end
-
-    def setup_rbs_spytest(modules, sdk_dir)
-      require 'rbs/test'
-      env = load_rbs_environment(sdk_dir)
-      tester = RBS::Test::Tester.new(env: env)
-
-      unchecked_classes = [
-        '::RSpec::Mocks::Double',
-        '::RSpec::Mocks::InstanceVerifyingDouble',
-        '::RSpec::Mocks::ObjectVerifyingDouble',
-        '::RSpec::Mocks::ClassVerifyingDouble'
-      ] # rubocop:disable
-
-      spy_modules = [modules.join('::'), 'Smithy::Client']
-      spy_classes = []
-      spy_modules.each do |module_name|
-        spy_classes += classes_to_spy(Object.const_get(module_name), env)
-      end
-
-      spy_classes.each do |spy_class|
-        tester.install!(
-          spy_class,
-          sample_size: RBS::Test::SetupHelper::DEFAULT_SAMPLE_SIZE,
-          unchecked_classes: unchecked_classes
-        )
-      end
     end
   end
 end
